add class

+++ CinemaManager/src/Phim.java
@@ -2,7 +2,7 @@ public class Phim{
     String maPhim;
     String tenPhim;
     String theLoai;
-    float thoiLuong;   // phút
+    int thoiLuong;   // đơn vị: phút
     String ngonNgu;
     int gioiHanTuoi;
     String moTa;
